fix bug

+++ app/post/postDetailController.js
@@ -5,17 +5,25 @@
     angular.module('MyApp')
         .controller('postDetailController', postDetailController);
 
-    postDetailController.$inject = ['$scope', 'postService', '$location', '$routeParams', 'headerService', '$window'];
+    postDetailController.$inject = ['$scope', 'ngDialog', 'postService', '$location', '$routeParams', 'headerService', '$window'];
 
-    function postDetailController($scope, postService, $location, $routeParam, headerService, $window) {
+    function postDetailController($scope,ngDialog, postService, $location, $routeParam, headerService, $window) {
         $scope.iconLike = false;
         $scope.iconFavorite = false;
         $scope.editing = false;
         $scope.editingComment = false;
-
         postService.getPost($routeParam.id)
             .success(function(data){
                 $scope.post = data.post;
+
+
+                $scope.captionCustom=$scope.post.caption.substr(0,48)+"\n";
+                for(var i=0;i<4;i++){
+                    $scope.captionCustom=$scope.captionCustom+$scope.post.caption.substr(0,48)+"\n";
+                }
+                $scope.captionCustom=$scope.captionCustom+$scope.post.caption.substr(240,$scope.post.caption.length-240);
+                $scope.post.caption=$scope.captionCustom;
+
                 $scope.post.created_at = beautyDate.prettyDate($scope.post.created_at);
                 for (var i=0; i< $scope.post.tag_picture.length; i++){
                     $scope.post.tag_picture[i].top = $scope.post.tag_picture[i].top - 20;
@@ -52,29 +60,31 @@
                         event.preventDefault();
                         headerService.openLogin();
                     } else {
-                        var data = {
-                            'content': this.comment,
-                            'type_comment': 0,
-                            'type_id': $routeParam.id,
-                            'user_id': $scope.loginUser.id
-                        };
-                        this.comment = null;
-                        //$scope.post.comments.push({
-                        //    'content': data.content,
-                        //    'created_at': 'Just now',
-                        //    'user': {
-                        //        'username': $scope.loginUser.username,
-                        //        'id': $scope.loginUser.id,
-                        //        'picture_profile': $scope.loginUser.picture_profile
-                        //    }
-                        //});
-
-                        postService.saveComment(data)
-                            .success(function (data) {
-                            })
-                            .error(function (data) {
-
+                        if(this.comment != null){
+                            var data = {
+                                'content': this.comment,
+                                'type_comment': 0,
+                                'type_id': $routeParam.id,
+                                'user_id': $scope.loginUser.id
+                            };
+                            this.comment = null;
+                            $scope.post.comments.push({
+                                'content': data.content,
+                                'created_at': 'Just now',
+                                'user': {
+                                    'username': $scope.loginUser.username,
+                                    'id': $scope.loginUser.id,
+                                    'picture_profile': $scope.loginUser.picture_profile
+                                }
                             });
+
+                            postService.saveComment(data)
+                                .success(function (data) {
+                                })
+                                .error(function (data) {
+
+                                });
+                        }
                     }
                 };
                 $scope.likeOrDislike = function () {
@@ -141,25 +151,47 @@
                     $scope.editContent = content;
                 };
 
-                $scope.submitEditComment = function (index) {
-                    $scope.post.comments[index].content = this.editContent;
-                    $scope.post.comments[index].editing = null;
-                    postService.editPostComment({
-                        id: $scope.post.comments[index].id,
-                        content: this.editContent
-                    }).success(function (data) {
+                $scope.submitEditComment = function (index,editContent) {
+                    if(editContent.length > 255 ){
+                        ngDialog.open({
+                            template: 'app/post/templates/alertInpTxtLength.html',
+                            className: 'ngdialog-theme-plain-custom',
 
-                    }).error();
+                            controller: ['$scope', 'postService', function ($scope, postService) {
+                            }]
+                        });
+                    }else{
+                        $scope.post.comments[index].content = this.editContent;
+                        $scope.post.comments[index].editing = null;
+                        postService.editPostComment({
+                            id: $scope.post.comments[index].id,
+                            content: this.editContent
+                        }).success(function (data) {
+
+                        }).error();
+                    }
                 };
 
-                $scope.deleteCommentIndex = function (index) {
+                $scope.deleteCommentIndex = function (index,comments) {
+                    ngDialog.open({
+                        template: 'app/post/templates/confirmDeletePost.html',
+                        className: 'ngdialog-theme-plain-custom',
 
-                    postService.deletePostComment({
-                        id: $scope.post.comments[index].id
-                    }).success(function (data) {
+                        controller: ['$scope', function ($scope) {
+                            $scope.close = function () {
+                                ngDialog.close();
+                            };
+                            $scope.confirm = function () {
+                                ngDialog.close();
+                                postService.deletePostComment({
+                                    id: comments[index].id
+                                }).success(function (data) {
 
-                    }).error();
-                    $scope.post.comments.splice(index, 1);
+                                }).error();
+                                comments.splice(index, 1);
+                            }
+                        }]
+                    });
                 };
 
                 $scope.closeEditComment = function (index) {
@@ -189,11 +221,24 @@
                 };
 
                 $scope.deletePostInside = function () {
-                    postService.delete({
-                        id: $routeParam.id
-                    }).success(function (data) {
-                        $window.location.reload();
-                    }).error();
+                    ngDialog.open({
+                        template: 'app/post/templates/confirmDeletePost.html',
+                        className: 'ngdialog-theme-plain-custom',
+
+                        controller: ['$scope', function ($scope) {
+                            $scope.close = function () {
+                                ngDialog.close();
+                            };
+                            $scope.confirm = function () {
+                                ngDialog.close();
+                                postService.delete({
+                                    id: $routeParam.id
+                                }).success(function (data) {
+                                    $window.location.reload();
+                                }).error();
+                            }
+                        }]
+                    });
                 };
 
                 $scope.closeDialog = function () {
